pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["alexey <a556077@icloud.com>"]
readme = "README.md"
packages = [{include = "poetry_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg = "^3.1.9"
psycopg-binary = "^3.1.9"
python-dotenv = "^1.0.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy-utils = "^0.41.1"
pydantic = "^2.5.3"
fastapi = "^0.108.0"
pydantic-settings = "^2.1.0"
alembic = "1.13.1"
sqlalchemy = "^2.0.25"
minio = "^7.2.3"
python-multipart = "^0.0.6"
fastapi-filter = "^1.1.0"
gunicorn = "21.2.0"
pillow = "^10.2.0"
numpy = "^1.26.3"
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

